requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
joblib==1.3.2
psycopg2-binary==2.9.10
pandas==2.0.3
xgboost==3.0.3
scikit-learn==1.3.0